Add rdoc comments, refactor and rename CreateBadgeService to AssignBadgeService
